+++ nukibridge.js
@@ -1,6 +1,7 @@
 var request = require("request");
 var http = require('http');
 var https = require('https');
+var url = require('url');
 
 global.NUKI_LOCK_ACTION_UNLOCK = "1";
 global.NUKI_LOCK_ACTION_LOCK = "2";
@@ -22,11 +23,12 @@ global.NUKI_LOCK_STATE_UNDEFINED = 255;
 var MORE_LOGGING = false;
 var DEFAULT_REQUEST_TIMEOUT_LOCK_STATE = 5000;
 var DEFAULT_REQUEST_TIMEOUT_LOCK_ACTION = 30000;
+var DEFAULT_WEBHOOK_SERVER_PORT = 51827;
 var DEFAULT_CACHE_DIRECTORY = "./.node-persist/storage";
 
 var instances = [];
 
-this.getInstance = function(constructorLog, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory) {
+this.getInstance = function(constructorLog, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory, webHookServerPort) {
     var instanceArray = instances.filter(function(nukiBridge){
         return nukiBridge.bridgeUrl == bridgeUrl;
     });
@@ -34,13 +36,13 @@ this.getInstance = function(constructorLog, bridgeUrl, apiToken, requestTimeoutL
         return instanceArray[0];
     }
     else {
-        var newBridge = new NukiBridge(constructorLog, instances.length, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory);
+        var newBridge = new NukiBridge(constructorLog, instances.length, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory, webHookServerPort);
         instances.push(newBridge);
         return newBridge;
     }
 };
 
-function NukiBridge(constructorLog, instanceId, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory) {
+function NukiBridge(constructorLog, instanceId, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory, webHookServerPort) {
     constructorLog("Initializing Nuki bridge '%s' (instance id '%s')...", bridgeUrl, instanceId);
     this.instanceId = instanceId;
     this.bridgeUrl = bridgeUrl;
@@ -48,6 +50,7 @@ function NukiBridge(constructorLog, instanceId, bridgeUrl, apiToken, requestTime
     this.requestTimeoutLockState = requestTimeoutLockState;
     this.requestTimeoutLockAction = requestTimeoutLockAction;
     this.cacheDirectory = cacheDirectory;
+    this.webHookServerPort = webHookServerPort;
     if(this.requestTimeoutLockState == null || this.requestTimeoutLockState == "" || this.requestTimeoutLockState < 1) {
         this.requestTimeoutLockState = DEFAULT_REQUEST_TIMEOUT_LOCK_STATE;
     }
@@ -57,6 +60,9 @@ function NukiBridge(constructorLog, instanceId, bridgeUrl, apiToken, requestTime
     if(this.cacheDirectory == null || this.cacheDirectory == "") {
         this.cacheDirectory = DEFAULT_CACHE_DIRECTORY;
     }
+    if(this.webHookServerPort == null || this.webHookServerPort == "") {
+        this.webHookServerPort = DEFAULT_WEBHOOK_SERVER_PORT;
+    }
     this.storage = require('node-persist');
     this.storage.initSync({dir:this.cacheDirectory});
     
@@ -64,12 +70,87 @@ function NukiBridge(constructorLog, instanceId, bridgeUrl, apiToken, requestTime
     this.queue = [];
     this.locks = [];
     
+    this._createWebHookServer(constructorLog, this.webHookServerPort);
+    
     constructorLog("Initialized Nuki bridge (instance id '%s').", instanceId);
+};
+
+NukiBridge.prototype._createWebHookServer = function _createWebHookServer(constructorLog, webHookServerPort) {
+    http.createServer((function(request, response) {
+        var theUrl = request.url;
+        var theUrlParts = url.parse(theUrl, true);
+        var theUrlParams = theUrlParts.query;
+        var body = [];
+        request.on('error', function(err) {
+            console.error("[ERROR Nuki WebHook Server] Reason: %s.", err);
+        }).on('data', function(chunk) {
+            body.push(chunk);
+        }).on('end', (function() {
+            body = Buffer.concat(body).toString();
+
+            response.on('error', function(err) {
+            console.error("[ERROR Nuki WebHook Server] Reason: %s.", err);
+            });
+
+            response.statusCode = 200;
+            response.setHeader('Content-Type', 'application/json');
+
+            if(!theUrlParams.nukiId || !theUrlParams.state) {
+                response.statusCode = 404;
+                response.setHeader("Content-Type", "text/plain");
+                var errorText = "[ERROR Nuki WebHook Server] No nukiId or state in request.";
+                console.error(errorText);
+                response.write(errorText);
+                response.end();
+            }
+            else {
+                var nukiId = theUrlParams.nukiId;
+                var state = theUrlParams.state;
+                var ignoreDoorsWithDoorLatches = true;
+                var lock = this._getLock(nukiId, ignoreDoorsWithDoorLatches);
+                if(lock == null) {
+                    response.statusCode = 404;
+                    response.setHeader("Content-Type", "text/plain");
+                    var errorText = "[ERROR Nuki WebHook Server] No lock found for nukiId '"+nukiId+"'.";
+                    console.error(errorText);
+                    response.write(errorText);
+                    response.end();
+                }
+                else {            
+                    var responseBody = {
+                        success: true
+                    };
+                    if(!lock.isDoorLatch()) {
+                        var isLocked = lock._isLocked(state);
+                        lock._setIsLockedCache(isLocked);  
+                        console.log("[INFO Nuki WebHook Server] Updated lock state from webhook to isLocked = '%s' (Nuki state '%s' ) for lock '%s'.", isLocked, state, lock.instanceId);
+                        lock.webHookCallback(isLocked);
+                    }                      
+                    response.write(JSON.stringify(responseBody));
+                    response.end();
+                }
+            }
+        }).bind(this));
+    }).bind(this)).listen(webHookServerPort);    
+    constructorLog("Started server for webhooks on port '%s'.", webHookServerPort);
 };
 
 NukiBridge.prototype.addLock = function addLock(nukiLock) {
     nukiLock.instanceId = this.locks.length;
     this.locks.push(nukiLock);
+};
+
+NukiBridge.prototype._getLock = function _getLock(lockId, ignoreDoorsWithDoorLatches) {
+    for (var i = 0; i < this.locks.length; i++) {
+        var lock = this.locks[i];
+        if(lock.isDoorLatch() && ignoreDoorsWithDoorLatches) {
+            continue;
+        }
+        if(lock.lockId === lockId) {
+            return lock;
+        }
+    }
+    return null;
 };
 
 NukiBridge.prototype.lockState = function lockState(nukiLock, callback /*(err, json)*/) {
@@ -107,9 +188,7 @@ NukiBridge.prototype.lockAction = function lockAction(nukiLock, lockAction, call
 };
 
 NukiBridge.prototype._sendRequest = function _sendRequest(nukiLock, entryPoint, queryObject, requestTimeout, callback /*(err, json)*/) {
-    if(MORE_LOGGING) {
-        nukiLock.log("Send request to Nuki bridge '%s' on '%s' with '%s'.", this.instanceId, entryPoint, JSON.stringify(queryObject));
-    }
+    nukiLock.log("Send request to Nuki bridge '%s' on '%s' with '%s'.", this.instanceId, entryPoint, JSON.stringify(queryObject));
     this.runningRequest = true;
     request.get({
         url: this.bridgeUrl+entryPoint,
@@ -168,12 +247,13 @@ NukiBridge.prototype._processNextQueueEntry = function _processNextQueueEntry() 
     }
 };
 
-function NukiLock(log, nukiBridge, lockId, lockAction, unlockAction) {
+function NukiLock(log, nukiBridge, lockId, lockAction, unlockAction, webHookCallback) {
     this.nukiBridge = nukiBridge;
     this.log = log;
     this.lockId = lockId;
     this.lockAction = lockAction;
     this.unlockAction = unlockAction;
+    this.webHookCallback = webHookCallback;
     
     if(this.lockAction == null || this.lockAction == "") {
         this.lockAction = NUKI_LOCK_ACTION_LOCK;
@@ -207,12 +287,7 @@ NukiLock.prototype.isLocked = function isLocked(callback /*(err, isLocked)*/) {
                 if(json) {
                     state = json.state;
                 }
-                var isLocked = 
-                    state == NUKI_LOCK_STATE_LOCKED || 
-                    state == NUKI_LOCK_STATE_LOCKING || 
-                    state == NUKI_LOCK_STATE_UNCALIBRATED || 
-                    state == NUKI_LOCK_STATE_MOTOR_BLOCKED || 
-                    state == NUKI_LOCK_STATE_UNDEFINED;
+                var isLocked = this._isLocked(state);
                 this.log("Lock state is isLocked = '%s' (Nuki state '%s' )", isLocked, state);
                 this._setIsLockedCache(isLocked);
                 callback(null, isLocked);
@@ -222,9 +297,25 @@ NukiLock.prototype.isLocked = function isLocked(callback /*(err, isLocked)*/) {
     }
 };
 
+NukiLock.prototype._isLocked = function _isLocked(state) {
+    var isLocked = 
+        state == NUKI_LOCK_STATE_LOCKED || 
+        state == NUKI_LOCK_STATE_LOCKING || 
+        state == NUKI_LOCK_STATE_UNCALIBRATED || 
+        state == NUKI_LOCK_STATE_MOTOR_BLOCKED || 
+        state == NUKI_LOCK_STATE_UNDEFINED;
+    return isLocked;
+};
+
 NukiLock.prototype.lock = function lock(callback) {
-    this._setIsLockedCache(true);
-    this.nukiBridge.lockAction(this, this.lockAction, callback);
+    if(this.isDoorLatch()) {
+        this.log("Doors with door latch never process action to lock as these locks are always locked.");
+        callback(null);
+    }
+    else{
+        this._setIsLockedCache(true);
+        this.nukiBridge.lockAction(this, this.lockAction, callback);
+    }
 };
 
 NukiLock.prototype.unlock = function unlock(callback) {